feat: remove rocketcode from move_app an create generic enry creation function
